nlp_service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
transformers==4.36.2
torch==2.1.2
spacy==3.7.2
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.1
python-multipart==0.0.6
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl 